Use SLF4J for all logging
In an attempt to get a more standardised method for logging, this patch
replaces all uses of nl.knmi.adaguc.tools.Debug, System.out, and
System.err with the use of SLF4J Loggers. Furthermore, this patch
configures Logback to keep the logging behaviour as close to that of
n.k.a.t.Debug as possible.